refactor: message manager 구조 개편
pros
빌더의 명시적인 분리 -> send로 매니저 생성
send할 때 message 대신 channel를 받게 하여 구조적으로 "waiting..." 과 같은 메시지가 안나오게 개선
여기서 실제 유즈케이스에선 별 차이를 못느끼게 바꿈 (사실 걍 'channel' | 'interaction' 타입의 인자만 더 넘기면 됨)

cons
send 호출 시 channel에 의한 send인지 interaction에 의한 send인지 명시적으로 구분이 필요
